fix: test code 수정

+++ server/server-application/src/main/java/com/mzdevelopers/serverapplication/answer/service/AnswerService.java
@@ -4,11 +4,10 @@ import com.mzdevelopers.serverapplication.answer.entity.Answer;
 import com.mzdevelopers.serverapplication.answer.repository.AnswerRepository;
 import com.mzdevelopers.serverapplication.answervote.entity.AnswerVote;
 import com.mzdevelopers.serverapplication.answervote.repository.AnswerVoteRepository;
+import com.mzdevelopers.serverapplication.member.entity.Member;
+import com.mzdevelopers.serverapplication.member.repository.MemberRepository;
 import com.mzdevelopers.serverapplication.question.entity.Question;
-import com.mzdevelopers.serverapplication.question.entity.QuestionVote;
 import com.mzdevelopers.serverapplication.question.repository.QuestionRepository;
-import com.mzdevelopers.serverapplication.question.stub.MemberStub;
-import com.mzdevelopers.serverapplication.question.stub.MemberStubRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -27,12 +26,14 @@ public class AnswerService {
     private final AnswerRepository answerRepository;
     private final QuestionRepository questionRepository;
     private final AnswerVoteRepository answerVoteRepository;
-    private final MemberStubRepository memberStubRepository;
+    private final MemberRepository memberRepository;
 
 
 
     public Answer createAnswer(Answer answer){
         Question findQuestion = questionRepository.findByQuestionId(answer.getQuestion().getQuestionId());
+        Member findMember = memberRepository.findById(answer.getMember().getId()).orElseThrow(()->new RuntimeException("멤버를 찾을 수 없습니다."));
+        answer.setMember(findMember);
 
         Answer saveAnswer = answerRepository.save(answer);
         findQuestion.getAnswers().add(saveAnswer);
@@ -44,6 +45,11 @@ public class AnswerService {
     @Transactional(propagation = Propagation.REQUIRED, isolation = Isolation.SERIALIZABLE)
     public Answer updateAnswer(Answer answer){
         Answer findAnswer = findVerifiedAnswer(answer.getAnswerId());
+        Member findMember = memberRepository.findById(answer.getMember().getId()).orElseThrow(()->new RuntimeException("멤버를 찾을 수 없습니다."));
+
+        if(findAnswer.getMember().getId()!=findMember.getId()){
+            throw new RuntimeException("수정할 권한이 없습니다.");
+        }
 
         Optional.ofNullable(answer.getDetail())
                 .ifPresent(detail -> findAnswer.setDetail(detail));
@@ -57,7 +63,7 @@ public class AnswerService {
     @Transactional(readOnly = true)
     public Answer findAnswer(long answerId){
 
-        return answerRepository.findByAnswerId(answerId);
+        return answerRepository.findByAnswerId(answerId).orElseGet(null);
     }
 
 
@@ -65,8 +71,14 @@ public class AnswerService {
         return answerRepository.findAll(PageRequest.of(page, size,
                 Sort.by("answerId").descending()));
     }
-    public void deleteAnswer(long answerId) {
+    public void deleteAnswer(long answerId, long memberId) {
         Answer findAnswer = findVerifiedAnswer(answerId);
+        Member findMember = memberRepository.findById(memberId).orElseThrow(()->new RuntimeException("멤버를 찾을 수 없습니다."));
+
+        if(findAnswer.getMember().getId()!=findMember.getId()){
+            throw new RuntimeException("삭제할 권한이 없습니다.");
+        }
+
 
         answerRepository.delete(findAnswer);
     }
@@ -75,17 +87,17 @@ public class AnswerService {
     @Transactional(readOnly = true)
     public Answer findVerifiedAnswer(long answerId){
         Optional<Answer> optionalAnswer =
-                Optional.ofNullable(answerRepository.findByAnswerId(answerId));
+                answerRepository.findByAnswerId(answerId);
         return optionalAnswer.orElseThrow();
     }
 
 
     public int votesCount(long answerId, long memberId) {
         Answer findAnswer = findByAnswerId(answerId);
-        MemberStub memberStub = save();
-        Optional<AnswerVote> optionalAnswerVote = answerVoteRepository.findByAnswerAndMemberStub(findAnswer, memberStub);
+        Member findMember = findByMemberId(memberId);
+        Optional<AnswerVote> optionalAnswerVote = answerVoteRepository.findByAnswerAndMember(findAnswer, findMember);
         if (optionalAnswerVote.isEmpty()) {
-            AnswerVote answerVote = AnswerVote.builder().answer(findAnswer).memberStub(memberStub).build();
+            AnswerVote answerVote = AnswerVote.builder().answer(findAnswer).member(findMember).build();
             findAnswer.updateVoteCount(true);
             answerVoteRepository.save(answerVote);
         } else {
@@ -103,9 +115,8 @@ public class AnswerService {
         return findAnswer.orElseThrow(() -> new IllegalArgumentException("No Search Question: " + answerId));
     }
 
-    public MemberStub save(){
-        MemberStub memberStub = MemberStub.builder().name("member").build();
-        memberStubRepository.save(memberStub);
-        return memberStubRepository.findById(1L).orElseGet(null);
+    public Member findByMemberId(Long memberId){
+        Optional<Member> findMember = memberRepository.findById(memberId);
+        return findMember.orElseThrow(() -> new IllegalArgumentException("No Search Member: " + memberId));
     }
 }
