+++ server/server-application/src/main/java/com/mzdevelopers/serverapplication/comment/service/CommentService.java
@@ -5,6 +5,8 @@ import com.mzdevelopers.serverapplication.answer.repository.AnswerRepository;
 import com.mzdevelopers.serverapplication.answer.service.AnswerService;
 import com.mzdevelopers.serverapplication.comment.entity.Comment;
 import com.mzdevelopers.serverapplication.comment.repository.CommentRepository;
+import com.mzdevelopers.serverapplication.member.entity.Member;
+import com.mzdevelopers.serverapplication.member.repository.MemberRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -20,17 +22,23 @@ import java.util.Optional;
 public class CommentService {
     private final CommentRepository commentRepository;
     private final AnswerRepository answerRepository;
+    private final MemberRepository memberRepository;
     private final AnswerService answerService;
 
-    public CommentService(CommentRepository commentRepository, AnswerRepository answerRepository, AnswerService answerService) {
+    public CommentService(CommentRepository commentRepository, AnswerRepository answerRepository, MemberRepository memberRepository, AnswerService answerService) {
         this.commentRepository = commentRepository;
         this.answerRepository = answerRepository;
+        this.memberRepository = memberRepository;
         this.answerService = answerService;
     }
 
     public Comment createComment(Comment comment){
         verifyComment(comment);
-        Answer findAnswer = answerRepository.findByAnswerId(comment.getAnswer().getAnswerId());
+        Answer findAnswer = answerRepository.findByAnswerId(comment.getAnswer().getAnswerId()).orElseThrow(()->new RuntimeException("회원을 찾지 못했습니다."));
+        Member findMember = memberRepository.findById(comment.getMember().getId()).orElseThrow(()->new RuntimeException("회원을 찾지 못했습니다."));
+
+        comment.setAnswer(findAnswer);
+        comment.setMember(findMember);
         Comment savecomment = commentRepository.save(comment);
         findAnswer.getComments().add(savecomment);
         answerRepository.save(findAnswer);
@@ -38,8 +46,13 @@ public class CommentService {
     }
 
     @Transactional(propagation = Propagation.REQUIRED, isolation = Isolation.SERIALIZABLE)
-    public Comment updateComment(Comment comment){
+    public Comment updateComment(Comment comment, long memberId){
         Comment findComment = findVerifiedComment(comment.getCommentId());
+        Member findMember = memberRepository.findById(memberId).orElseThrow(()->new RuntimeException("회원을 찾지 못했습니다."));
+
+        if(findMember.getId()!=findComment.getMember().getId()){
+            throw new RuntimeException("수정할 권한이 없습니다.");
+        }
 
         Optional.ofNullable(comment.getCommentDetail())
                 .ifPresent(detail -> findComment.setCommentDetail(detail));
@@ -47,6 +60,7 @@ public class CommentService {
 
         return commentRepository.save(findComment);
     }
+
     @Transactional(readOnly = true)
     public Comment findComment(long commentId){
         return findVerifiedComment(commentId);
@@ -57,8 +71,14 @@ public class CommentService {
                 Sort.by("commentId").descending()));
     }
 
-    public void deleteComment(long commentId) {
+    public void deleteComment(long commentId, long memberId) {
         Comment findComment = findVerifiedComment(commentId);
+        Member findMember = memberRepository.findById(memberId).orElseThrow(()->new RuntimeException("회원을 찾지 못했습니다."));
+
+        if(findMember.getId()!=findComment.getMember().getId()){
+            throw new RuntimeException("삭제할 권한이 없습니다.");
+        }
+
 
         commentRepository.delete(findComment);
     }
